+++ src/Flights.Data/Models/Result/FlightSolution.cs
@@ -3,7 +3,7 @@ namespace Flights.Data.Models.Result;
 public class FlightSolution
 {
     public required string Id { get; set; }
-    public required decimal Price { get; set; }
+    public required double Price { get; set; }
     public int PassengerCount { get; set; }
     public required FlightSlice[] Slices { get; set; }
 }
